requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>0.3.0
python-dotenv>=1.0.0
duckduckgo-search>=7.0.0
beautifulsoup4>=4.12.2
html2text>=2024.2.26
